\name{SVCJP-package}
\alias{SVCJP-package}
\alias{SVCJP}
\docType{package}
\title{
\packageTitle{SVCJP}
}
\description{
\packageDescription{SVCJP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SVCJP}
\packageIndices{SVCJP}
}
\author{
\packageAuthor{SVCJP}

Maintainer: \packageMaintainer{SVCJP}
}
\references{
~~ Literature or other references for background information ~~
}